requirements.txt
fastapi==0.104.0
uvicorn[standard]==0.23.2
python-jose[cryptography]
passlib[bcrypt]
python-jose[cryptography]
python-multipart
redis
ujson
requests
SQLAlchemy
psycopg2
pandas